+++ cmake/ExternalDependencies.cmake
@@ -67,11 +67,11 @@ if(BUILD_MQT_CORE_TESTS)
 endif()
 
 # cmake-format: off
-set(QDMI_MINIMUM_VERSION 1.3.3
+set(QDMI_MINIMUM_VERSION 1.4
         CACHE STRING "Minimum QDMI version")
-set(QDMI_VERSION 1.3.3
+set(QDMI_VERSION 1.4.0
         CACHE STRING "QDMI version")
-set(QDMI_REV "18cfb67fd9042761d3005c2f8655751c1758f9c5" # v1.3.3
+set(QDMI_REV "4949ecba4b61e9492160b1be96507a570bea3765" # enum-based multi-program jobs
         CACHE STRING "QDMI identifier (tag, branch or commit hash)")
 set(QDMI_REPO_OWNER "Munich-Quantum-Software-Stack"
         CACHE STRING "QDMI repository owner (change when using a fork)")